finish texture test, update plan
texture now can now be showed on the cube based on the uv of each quad  which contains a uv position and a texture index.

Next, I am going to integrate and polish the existing scripts into usable model and utilities.

The terrain generation is another really big part of the Terra system, it need more abstraction and rules to build a outline for the costume terrain genration.